fix: the Await bridge moves to a PRODUCT assembly — product code deadlocks too
ObservableAwait (faithful ToTask semantics, but RunContinuationsAsynchronously so the continuation
is QUEUED rather than resumed inline on the signalling thread) lived in MeshWeaver.Fixture — a TEST
assembly. So the bridges that actually deadlock a hub had nowhere to go: an SDK boundary we must
implement (a Microsoft.Agent.Framework ...Async override, IHostedService, ASP.NET middleware) has a
signature that is not ours to change, cannot reference the test fixture, and was therefore left on
Rx's .ToTask().

Measured: core src/ carries ZERO .ToTask( and has a ratchet holding it there; MeshWeaver.Plugins
carries 65 in product code with no ratchet, 7 of them in MeshNodeAgentFileStore — the class whose
test produced the AgentStore=1 dirty-teardown residual under load.

Moved to MeshWeaver.Messaging.Hub beside ReactiveCompletion (the other sanctioned bridge).
MeshWeaver.Fixture keeps the name and forwards, so ~1,500 test call sites need no churn and there
is exactly ONE implementation of the semantics the sweep depends on.
